requirements.txt
torch
torchvision
torchaudio
streamlit
albumentations
opencv-python-headless
pillow
numpy
